bot/src/requirements.txt
# Flask
flask
# ---

# Test
pytest
# ---

# Database (3.7.0)
tinydb
# ---

# Google App Engine
gunicorn
# ---

# EasyNido gear
beautifulsoup4
requests
# ---

# Weather gear
weather-api
# ---

# Telegram support
urllib3
telepot
# ---

# Misc
# To parse json files for configuration, tasks etc
json_minify
# For date and time management
arrow
# ---

